:recycle: refactor: extract trace at chat()
