+++ src/app/api/chat/[provider]/route.ts
@@ -5,7 +5,7 @@ import { ChatErrorType } from '@/types/fetch';
 import { ChatStreamPayload } from '@/types/openai/chat';
 import { getTracePayload } from '@/utils/trace';
 
-import { initializeWithUserPayload } from '../agentRuntime';
+import { createTraceOptions, initializeWithUserPayload } from '../agentRuntime';
 import { checkAuth } from '../auth';
 
 export const runtime = 'edge';
@@ -25,11 +25,18 @@ export const POST = checkAuth(async (req: Request, { params, jwtPayload }) => {
 
     const tracePayload = getTracePayload(req);
 
-    return await agentRuntime.chat(data, {
-      enableTrace: tracePayload?.enabled,
-      provider,
-      trace: tracePayload,
-    });
+    // If user enable trace
+    if (tracePayload?.enabled) {
+      return await agentRuntime.chat(
+        data,
+        createTraceOptions(data, {
+          provider,
+          trace: tracePayload,
+        }),
+      );
+    }
+
+    return await agentRuntime.chat(data);
   } catch (e) {
     const {
       errorType = ChatErrorType.InternalServerError,
