+++ src/libs/agent-runtime/AgentRuntime.ts
@@ -1,13 +1,8 @@
 import { ClientOptions } from 'openai';
 
-import { INBOX_SESSION_ID } from '@/const/session';
+import type { TracePayload } from '@/const/trace';
 import {
-  LOBE_CHAT_OBSERVATION_ID,
-  LOBE_CHAT_TRACE_ID,
-  TracePayload,
-  TraceTagMap,
-} from '@/const/trace';
-import {
+  ChatCompetitionOptions,
   ChatStreamPayload,
   LobeAnthropicAI,
   LobeAzureOpenAI,
@@ -27,7 +22,6 @@ import {
   ModelProvider,
 } from '@/libs/agent-runtime';
 import { LobeBedrockAIParams } from '@/libs/agent-runtime/bedrock';
-import { TraceClient } from '@/libs/traces';
 
 export interface AgentChatOptions {
   enableTrace?: boolean;
@@ -42,67 +36,37 @@ class AgentRuntime {
     this._runtime = runtime;
   }
 
-  async chat(
-    payload: ChatStreamPayload,
-    { trace: tracePayload, provider, enableTrace }: AgentChatOptions,
-  ) {
-    const { messages, model, tools, ...parameters } = payload;
-
-    // if not enabled trace then just call the runtime
-    if (!enableTrace) return this._runtime.chat(payload);
-
-    // create a trace to monitor the completion
-    const traceClient = new TraceClient();
-    const trace = traceClient.createTrace({
-      id: tracePayload?.traceId,
-      input: messages,
-      metadata: { provider },
-      name: tracePayload?.traceName,
-      sessionId: `${tracePayload?.sessionId || INBOX_SESSION_ID}@${tracePayload?.topicId || 'start'}`,
-      tags: tracePayload?.tags,
-      userId: tracePayload?.userId,
-    });
-
-    const generation = trace?.generation({
-      input: messages,
-      metadata: { provider },
-      model,
-      modelParameters: parameters as any,
-      name: `Chat Completion (${provider})`,
-      startTime: new Date(),
-    });
-
-    return this._runtime.chat(payload, {
-      callback: {
-        experimental_onToolCall: async () => {
-          trace?.update({
-            tags: [...(tracePayload?.tags || []), TraceTagMap.ToolsCall],
-          });
-        },
-
-        onCompletion: async (completion) => {
-          generation?.update({
-            endTime: new Date(),
-            metadata: { provider, tools },
-            output: completion,
-          });
-
-          trace?.update({ output: completion });
-        },
-
-        onFinal: async () => {
-          await traceClient.shutdownAsync();
-        },
-
-        onStart: () => {
-          generation?.update({ completionStartTime: new Date() });
-        },
-      },
-      headers: {
-        [LOBE_CHAT_OBSERVATION_ID]: generation?.id,
-        [LOBE_CHAT_TRACE_ID]: trace?.id,
-      },
-    });
+  /**
+   * Initiates a chat session with the agent.
+   *
+   * @param payload - The payload containing the chat stream data.
+   * @param options - Optional chat competition options.
+   * @returns A Promise that resolves to the chat response.
+   *
+   * @example - Use without trace
+   * ```ts
+   * const agentRuntime = await initializeWithClientStore(provider, payload);
+   * const data = payload as ChatStreamPayload;
+   * return await agentRuntime.chat(data);
+   * ```
+   *
+   * @example - Use Langfuse trace
+   * ```ts
+   * // ============  1. init chat model   ============ //
+   * const agentRuntime = await initializeWithUserPayload(provider, jwtPayload);
+   * // ============  2. create chat completion   ============ //
+   * const data = {
+   * // your trace options here
+   *  } as ChatStreamPayload;
+   * const tracePayload = getTracePayload(req);
+   * return await agentRuntime.chat(data, createTraceOptions(data, {
+   *   provider,
+   *   trace: tracePayload,
+   * }));
+   * ```
+   */
+  async chat(payload: ChatStreamPayload, options?: ChatCompetitionOptions) {
+    return this._runtime.chat(payload, options);
   }
 
   async models() {
